+++ src/fluids/cosmicrays/multigrid_diffusion.F90
@@ -633,7 +633,7 @@ contains
 
       integer(kind=4),              intent(in)  :: crdim        !< direction in which we calculate flux
       integer, dimension(ndims),    intent(in)  :: im           !< [first cell index, second cell index, third cell index]
-      integer(kind=4),              intent(in)  :: soln         !< multigrid variable to differentiate
+      integer,                                   intent(in)  :: soln         !< multigrid variable to differentiate
       type(cg_list_level), pointer, intent(in)  :: curl         !< level on which differentiate
       integer,                      intent(in)  :: cr_id        !< CR component index
       real, optional,               intent(out) :: Keff         !< effective diffusion coefficient for relaxation
@@ -710,10 +710,10 @@ contains
       implicit none
 
       type(cg_list_level), pointer, intent(in) :: curl !< level for which approximate the solution
-      integer(kind=4),     intent(in) :: src   !< index of source in cg%q(:)
-      integer(kind=4),     intent(in) :: soln  !< index of solution in cg%q(:)
-      integer(kind=4),     intent(in) :: def   !< index of defect in cg%q(:)
-      integer,             intent(in) :: cr_id !< CR component index
+      integer,                      intent(in) :: src   !< index of source in cg%q(:)
+      integer,                      intent(in) :: soln  !< index of solution in cg%q(:)
+      integer,                      intent(in) :: def   !< index of defect in cg%q(:)
+      integer,                      intent(in) :: cr_id !< CR component index
 
       integer                         :: i, j, k
       integer(kind=4)                 :: idir
@@ -777,9 +777,9 @@ contains
       implicit none
 
       type(cg_list_level), pointer, intent(in) :: curl  !< level for which approximate the solution
-      integer(kind=4),     intent(in) :: src   !< index of source in cg%q(:)
-      integer(kind=4),     intent(in) :: soln  !< index of solution in cg%q(:)
-      integer,             intent(in) :: cr_id !< CR component index
+      integer,                      intent(in) :: src   !< index of source in cg%q(:)
+      integer,                      intent(in) :: soln  !< index of solution in cg%q(:)
+      integer,                      intent(in) :: cr_id !< CR component index
 
       integer, parameter              :: RED_BLACK = 2 !< the checkerboard requires two sweeps
